putty: fix conversion warnings

+++ libs/putty/logging.c
@@ -68,7 +68,7 @@ static void logprintf(struct LogContext *ctx, const char *fmt, ...)
     data = dupvprintf(fmt, ap);
     va_end(ap);
 
-    logwrite(ctx, data, strlen(data));
+    logwrite(ctx, data, (int)strlen(data));
     sfree(data);
 }
 
@@ -136,7 +136,7 @@ static void logfopen_callback(void *handle, int mode)
          * output, because the whole point is that we haven't managed
          * to open any such log file :-)
          */
-        from_backend(ctx->frontend, 1, event, strlen(event));
+        from_backend(ctx->frontend, 1, event, (int)strlen(event));
         from_backend(ctx->frontend, 1, "\r\n", 2);
     }
     sfree(event);
@@ -360,7 +360,7 @@ void log_packet(void *handle, int direction, int type,
 	if (((p % 16) == 0) || (p == len) || omitted) {
 	    if (output_pos) {
 		strcpy(dumpdata + 10+1+3*16+2+output_pos, "\r\n");
-		logwrite(ctx, dumpdata, strlen(dumpdata));
+		logwrite(ctx, dumpdata, (int)strlen(dumpdata));
 		output_pos = 0;
 	    }
 	}
@@ -434,9 +434,9 @@ static Filename *xlatlognam(Filename *src, char *hostname, int port,
                             struct tm *tm)
 {
     char buf[32], *bufp;
-    int size;
+    size_t size;
     char *buffer;
-    int buflen, bufsize;
+    size_t buflen, bufsize;
     const char *s;
     Filename *ret;
