+++ libs/putty/misc.c
@@ -265,7 +265,7 @@ void prompt_ensure_result_size(prompt_t *pr, int newlen)
 }
 void prompt_set_result(prompt_t *pr, const char *newstr)
 {
-    prompt_ensure_result_size(pr, strlen(newstr) + 1);
+    prompt_ensure_result_size(pr, (int)strlen(newstr) + 1);
     strcpy(pr->result, newstr);
 }
 void free_prompts(prompts_t *p)
@@ -292,7 +292,7 @@ char *dupstr(const char *s)
 {
     char *p = NULL;
     if (s) {
-        int len = strlen(s);
+        int len = (int)strlen(s);
         p = snewn(len + 1, char);
         strcpy(p, s);
     }
@@ -302,7 +302,7 @@ char *dupstr(const char *s)
 /* Allocate the concatenation of N strings. Terminate arg list with NULL. */
 char *dupcat(const char *s1, ...)
 {
-    int len;
+    size_t len;
     char *p, *q, *sn;
     va_list ap;
 
@@ -501,7 +501,7 @@ char *strbuf_to_str(strbuf *buf)
 void strbuf_catfv(strbuf *buf, const char *fmt, va_list ap)
 {
     buf->s = dupvprintf_inner(buf->s, buf->len, &buf->size, fmt, ap);
-    buf->len += strlen(buf->s + buf->len);
+    buf->len += (int)strlen(buf->s + buf->len);
 }
 void strbuf_catf(strbuf *buf, const char *fmt, ...)
 {
@@ -518,8 +518,8 @@ void strbuf_catf(strbuf *buf, const char *fmt, ...)
 char *fgetline(FILE *fp)
 {
     char *ret = snewn(512, char);
-    int size = 512, len = 0;
-    while (fgets(ret + len, size - len, fp)) {
+    size_t size = 512, len = 0;
+    while (fgets(ret + len, (int)(size - len), fp)) {
 	len += strlen(ret + len);
 	if (len > 0 && ret[len-1] == '\n')
 	    break;		       /* got a newline, we're done */
@@ -544,7 +544,7 @@ char *fgetline(FILE *fp)
 char *chomp(char *str)
 {
     if (str) {
-        int len = strlen(str);
+        size_t len = strlen(str);
         while (len > 0 && (str[len-1] == '\r' || str[len-1] == '\n'))
             len--;
         str[len] = '\0';
@@ -674,24 +674,24 @@ int bufchain_size(bufchain *ch)
     return ch->buffersize;
 }
 
-void bufchain_add(bufchain *ch, const void *data, int len)
+void bufchain_add(bufchain *ch, const void *data, size_t len)
 {
     const char *buf = (const char *)data;
 
     if (len == 0) return;
 
-    ch->buffersize += len;
+    ch->buffersize += (int)len;
 
     while (len > 0) {
 	if (ch->tail && ch->tail->bufend < ch->tail->bufmax) {
-	    int copylen = min(len, ch->tail->bufmax - ch->tail->bufend);
+			size_t copylen = min(len, ch->tail->bufmax - ch->tail->bufend);
 	    memcpy(ch->tail->bufend, buf, copylen);
 	    buf += copylen;
 	    len -= copylen;
 	    ch->tail->bufend += copylen;
 	}
 	if (len > 0) {
-	    int grainlen =
+			size_t grainlen =
 		max(sizeof(struct bufchain_granule) + len, BUFFER_MIN_GRANULE);
 	    struct bufchain_granule *newbuf;
 	    newbuf = smalloc(grainlen);
@@ -714,10 +714,10 @@ void bufchain_consume(bufchain *ch, int len)
 
     assert(ch->buffersize >= len);
     while (len > 0) {
-	int remlen = len;
+  int remlen = len;
 	assert(ch->head != NULL);
 	if (remlen >= ch->head->bufend - ch->head->bufpos) {
-	    remlen = ch->head->bufend - ch->head->bufpos;
+			remlen = (int)(ch->head->bufend - ch->head->bufpos);
 	    tmp = ch->head;
 	    ch->head = tmp->next;
 	    if (!ch->head)
@@ -732,7 +732,7 @@ void bufchain_consume(bufchain *ch, int len)
 
 void bufchain_prefix(bufchain *ch, void **data, int *len)
 {
-    *len = ch->head->bufend - ch->head->bufpos;
+    *len = (int)(ch->head->bufend - ch->head->bufpos);
     *data = ch->head->bufpos;
 }
 
@@ -745,11 +745,11 @@ void bufchain_fetch(bufchain *ch, void *data, int len)
 
     assert(ch->buffersize >= len);
     while (len > 0) {
-	int remlen = len;
+      int remlen = len;
 
 	assert(tmp != NULL);
 	if (remlen >= tmp->bufend - tmp->bufpos)
-	    remlen = tmp->bufend - tmp->bufpos;
+			remlen = (int)(tmp->bufend - tmp->bufpos);
 	memcpy(data_c, tmp->bufpos, remlen);
 
 	tmp = tmp->next;
@@ -1121,7 +1121,7 @@ int smemeq(const void *av, const void *bv, size_t len)
 
 int match_ssh_id(int stringlen, const void *string, const char *id)
 {
-    int idlen = strlen(id);
+    size_t idlen = strlen(id);
     return (idlen == stringlen && !memcmp(string, id, idlen));
 }
